+++ packages/benchmark-js/src/options/Settings.ts
@@ -3,61 +3,33 @@ import { BenchmarkJobSettings } from '../types';
 import { _Nanosecond } from '../types.internal';
 
 export class Settings {
-    private static minTime: _Nanosecond = Time.ns(Math.max(Time.minResolution * 100, 50_000_000));
-
     private _delay: _Nanosecond;
-    private _initCount: number;
-    private _maxAdjustTime: _Nanosecond;
-    private _maxPreparingTime: _Nanosecond;
-    private _maxTime: _Nanosecond;
-    private _minSamples: number;
-    private _minTime: _Nanosecond;
+    private _initOpsCount: number;
+    private _minSampleTime: _Nanosecond;
+    private _samplesCount: number;
 
     public constructor(settings: BenchmarkJobSettings) {
-        const {
-            delay = 5,
-            initCount = 1,
-            maxAdjustTime = 100,
-            maxPreparingTime = 10,
-            maxTime = 5_000,
-            minSamples = 5,
-            minTime = 0,
-        } = settings;
+        const { delay = 5, initOpsCount = 1, minSampleTime = 100, samplesCount = 15 } = settings;
 
         this._delay = Time.ms2ns(delay);
-        this._initCount = initCount;
-        this._maxAdjustTime = Time.ms2ns(maxAdjustTime);
-        this._maxPreparingTime = Time.ms2ns(maxPreparingTime);
-        this._maxTime = Time.ms2ns(maxTime);
-        this._minSamples = minSamples;
-        this._minTime = minTime === 0 ? Settings.minTime : Time.ms2ns(minTime);
+        this._initOpsCount = initOpsCount;
+        this._minSampleTime = Time.ms2ns(minSampleTime);
+        this._samplesCount = samplesCount;
     }
 
     public get delay(): _Nanosecond {
         return this._delay;
     }
 
-    public get initCount(): number {
-        return this._initCount;
+    public get initOpsCount(): number {
+        return this._initOpsCount;
     }
 
-    public get maxAdjustTime(): _Nanosecond {
-        return this._maxAdjustTime;
+    public get minSampleTime(): _Nanosecond {
+        return this._minSampleTime;
     }
 
-    public get maxPreparingTime(): _Nanosecond {
-        return this._maxPreparingTime;
-    }
-
-    public get maxTime(): _Nanosecond {
-        return this._maxTime;
-    }
-
-    public get minSamples(): number {
-        return this._minSamples;
-    }
-
-    public get minTime(): _Nanosecond {
-        return this._minTime;
+    public get samplesCount(): number {
+        return this._samplesCount;
     }
 }