+++ packages/benchmark-js/CHANGELOG.md
@@ -14,6 +14,8 @@
   - `BenchmarkSettings`         -> `BenchmarkJobSettings`
   - `BenchmarkTestFnOptions`    -> `BenchmarkJobTestFnOptions`
 - Delete methods `Benchmark.writeTestersCode` and `BenchmarkJob.writeTesterCode`
+- Change stages to Jitting, Pilot and Formal
+- Change `BenchmarkJobOptions`
 
 ## @easy/benchmark-js v0.3.0 (2022-02-01)
 ### Features